refactor: split up server.go
as it's getting a bit unwieldy
