+++ internal/server/redirect_handler.go
@@ -1,0 +1,25 @@
+// Copyright 2024 Michael Davis
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package server
+
+import "net/http"
+
+type httpRedirectHandler struct {
+	protoAndHost string
+}
+
+func (h httpRedirectHandler) ServeHTTP(w http.ResponseWriter, req *http.Request) {
+	http.Redirect(w, req, h.protoAndHost+req.RequestURI, http.StatusMovedPermanently)
+}
